+++ mizani/palettes.py
@@ -445,12 +445,19 @@ def ratios_to_colors(values, colormap):
     out : list | float
         Color(s) corresponding to the values
     """
+    iterable = True
+    try:
+        iter(values)
+    except TypeError:
+        iterable = False
+        values = [values]
+
     color_tuples = colormap(values)
     try:
         hex_colors = [mcolors.rgb2hex(t) for t in color_tuples]
     except IndexError:
         hex_colors = mcolors.rgb2hex(color_tuples)
-    return hex_colors
+    return hex_colors if iterable else hex_colors[0]
 
 
 def gradient_n_pal(colors, values=None, name='gradientn'):